pyproject.toml
[tool.poetry]
name = "caption-generator-api"
version = "0.1.0"
description = ""
authors = ["Abduraxmonnn <abduraxmonasatullayev35@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
django-filter = "^24.3"
drf-yasg = "^1.21.8"
django-allauth = "^65.1.0"
boto3 = "^1.35.54"
requests = "^2.32.3"
assemblyai = "^0.35.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
